devices: virtio: iommu: narrow the pci range of viot
Now we can use option `--pcie-root-port` to create a virtual pcie root
port with some attributes same as host. So a vfio device may be
connected to it statically.

And only hot-pluggable devices and vfio devices with option
`iommu=viommu` are needed to be attached to virtio-iommu. So we just
append the sub pci range of a leaf node root port to viot.

BUG=b:185084350
TEST=Boot a guest with option '--virtio-iommu' and '--pcie-root-port',
check the PCI range node in guest viot.

Change-Id: Iff59b5210fb85511370ede19bf1ca2095b4d603b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3541469
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>