devices: vhost-user: handler: define trait for platform-dependent ops
The DeviceRequestHandler used to differenciate between regular
vhost-user and VVU using an internal enum. This doesn't scale and makes
platform-dependent code visible in the handler code.

Improve this by introducing a VhostUserPlatformOps trait that is
implemented by vhost-user and VVU supporting structs. This will allow us
to move these structs into a sys directory in the future, and to add
support for new platforms without it being intrusive.

As a nice side-effect, the trait can also be used for free in the vsock
device, removing some technical debt there.

The Doorbell also happens to be platform specific, so put it behind the
platform module too.

BUG=b:229554679
BUG=b:230665747
TEST=vhost-user console device works.
TEST=VVU console device works.
TEST=vhost-user vsock device works (tested with socat on both ends).
TEST=VVU vsock device works (tested with socat on both ends).

Change-Id: I8d9faece7e5631921e0b622efaafba5b7fa2df8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3599230
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>